luôn đọc gpio

+++ usb_to_com/Core/Inc/drive.h
@@ -85,7 +85,7 @@ void Robot_Init(void);
 void  MC_Control_Interrupt(void);
 uint8_t MC_MoveLinear(int32_t posx,int32_t posy,int32_t posz );// thời điểm kết thúc gần bằng nhau tuyệt đối
 void MC_MoveAbsoluteTest(uint32_t posx,uint32_t posy,uint32_t posz, uint32_t freq);
-void MC_MoveHandle(uint8_t axis,uint8_t status, int dir);
+void MC_MoveHandle(uint8_t axis,uint8_t status, uint8_t dir);
 uint8_t Move_Home_3Step(volatile uint8_t * home_tep);
 void Reset_position(void);
 void Emergency_Stop(void);
